auth

+++ app/Http/Controllers/DosenController.php
@@ -4,19 +4,44 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\dosen;
-use Input;
+use Auth;
+use App\User;
+use Exception;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DosenController extends Controller
 {
     public function index()
-    {
-    	$dosen = dosen::get();
-    	
-        return response()->json($dosen->toArray());
+    {   
+        $user=Auth::user();
+
+        if($user->id_level==3){
+
+            return response()->json(['error'=>Auth::user()->name.',Forbidden'], 403);
+
+        }elseif ($user->id_level ==1) {
+
+            $dosen = dosen::get();
+
+            return response()->json($dosen->toArray());
+
+        }elseif ($user->id_level==2){
+
+           $dosen = dosen::find($id);
+
+            return response()->json($dosen->toArray());
+        }
     }
 
      public function store(Request $request)
     {
+        $user=Auth::user();
+
+        if($user->id_level==3){
+
+            return response()->json(['error'=>Auth::user()->name.',Forbidden'], 403);
+
+        }elseif ($user->id_level ==1) {
 
         try{
             $this->validate($request, [
@@ -32,54 +57,89 @@ class DosenController extends Controller
             } 
 
         $dosen = new dosen();
-
-        $dosen->dosen = $request->input('nama');
-        $dosen->dosen = $request->input('tempat_lahir');
-        $dosen->dosen = $request->input('tanggal_lahir');
-        $dosen->dosen = $request->input('gender');
-        $dosen->dosen = $request->input('phone');
-        $dosen->dosen = $request->input('email');
+        $dosen->nama=$request->input('nama');
+        $dosen->tempat_lahir=$request->input('tempat_lahir');
+        $dosen->tanggal_lahir=$request->input('tanggal_lahir');
+        $dosen->gender=$request->input('gender');
+        $dosen->phone=$request->input('phone');
+        $dosen->email=$request->input('email');
         $dosen->save();
 
         return response()->json($dosen);
+    }elseif ($user->id_level==2){
+
+           $dosen = dosen::find($nid);
+
+            return response()->json($dosen->toArray());
+        }
     }
 
     public function show(Request $request)
     {
-    	$dosen = dosen::find($nid);
+    	$dosen = dosen::find($id);
 
         return response()->json($dosen->toArray());
 
     }
 
 
-    public function update(Request $request, $nid)
+    public function update(Request $request, $id)
     {
-    	$this->validate($request, [
-    		'nama'=>'required',
-    		'tempat_lahir'=>'required',
-    		'tanggal_lahir'=>'required',
-    		'gender'=>'required',
-    		'phone'=>'required',
-    		'email'=>'required',]);
-    	$dosen = dosen::find($nid);
-        $dosen->dosen = $request->input('nama');
-        $dosen->dosen = $request->input('tempat_lahir');
-        $dosen->dosen = $request->input('tanggal_lahir');
-        $dosen->dosen = $request->input('gender');
-        $dosen->dosen = $request->input('phone');
-        $dosen->dosen = $request->input('email');
+        $user=Auth::user();
 
-        $dosen->save();
+        if($user->id_level==3){
 
-        return response()->json($dosen->toArray());
+            return response()->json(['error'=>Auth::user()->name.',Forbidden'], 403);
+
+        }elseif ($user->id_level ==1) {
+
+        	$this->validate($request, [
+        		'nama'=>'required',
+        		'tempat_lahir'=>'required',
+        		'tanggal_lahir'=>'required',
+        		'gender'=>'required',
+        		'phone'=>'required',
+        		'email'=>'required',
+                ]);
+        	$dosen = dosen::find($id);
+            $dosen->nama = $request->input('nama');
+            $dosen->tempat_lahir = $request->input('tempat_lahir');
+            $dosen->tanggal_lahir = $request->input('tanggal_lahir');
+            $dosen->gender = $request->input('gender');
+            $dosen->phone = $request->input('phone');
+            $dosen->email = $request->input('email');
+            $dosen->save();
+
+            return response()->json($dosen->toArray());
+        
+        }elseif ($user->id_level==2){
+
+           $dosen = dosen::find($id);
+
+            return response()->json($dosen->toArray());
+        }
     }
 
-    public function destroy($nid)
+    public function destroy($id)
     {
-        $dosen = dosen::find($nid);
-        $dosen->delete();
+        $user=Auth::user();
 
-        return response()->json($dosen->toArray());
+        if($user->id_level==3){
+
+            return response()->json(['error'=>Auth::user()->name.',Forbidden'], 403);
+
+        }elseif ($user->id_level ==1) {
+
+            $dosen = dosen::find($id);
+            $dosen->delete();
+
+            return response()->json($dosen->toArray());
+
+        }elseif ($user->id_level==2){
+
+           $dosen = dosen::find($id);
+
+            return response()->json($dosen->toArray());
+        }
     }
 }
